MyString/src/MergeSortWithInversionPairs.java
import java.util.Arrays;

public class MergeSortWithInversionPairs {
	static int count = 0;

	public static void sort(Comparable a[]) {

		Comparable[] aux = new Comparable[a.length];
		sort(a, aux, 0, a.length - 1);
	}

	private static void sort(Comparable a[], Comparable aux[], int lo, int hi) {
		if (lo >= hi) {
			return;
		}
		int mid = lo + (hi - lo) / 2;
		sort(a, aux, lo, mid);
		sort(a, aux, mid + 1, hi);
		merge(a, aux, lo, mid, hi);
		count++;
		System.out
				.println("Merge Call " + count + " done: array is now" + Arrays.toString(a));
	}

	private static void merge(Comparable[] a, Comparable[] aux, int lo,
			int mid, int hi) {

		int i = lo;
		int j = mid + 1;

		for (int k = lo; k <= hi; k++) {
			aux[k] = a[k];
		}

		for (int k = lo; k <= hi; k++) {

			if (i > mid) {
				a[k] = aux[j++];
			} else if (j > hi) {
				a[k] = aux[i++];
			} else if (less(aux[j], aux[i])) {
				a[k] = aux[j++];
			} else {
				a[k] = aux[i++];
			}
		}

	}

	private static boolean less(Comparable v, Comparable w) {
		return v.compareTo(w) < 0;
	}

}
